+++ dwds/lib/src/services/chrome_proxy_service.dart
@@ -204,13 +204,9 @@ class ChromeProxyService implements VmServiceInterface {
     final moduleFormat = loadStrategy.moduleFormat;
     final canaryFeatures = loadStrategy.buildSettings.canaryFeatures;
     final experiments = loadStrategy.buildSettings.experiments;
+    final soundNullSafety = true;
 
-    // TODO(annagrin): Read null safety setting from the build settings.
-    final metadataProvider = loadStrategy.metadataProviderFor(entrypoint);
-    final soundNullSafety = await metadataProvider.soundNullSafety;
-
-    _logger.info('Initializing expression compiler for $entrypoint '
-        'with sound null safety: $soundNullSafety');
+    _logger.info('Initializing expression compiler for $entrypoint');
 
     final compilerOptions = CompilerOptions(
       moduleFormat: ModuleFormat.values.byName(moduleFormat),