+++ dwds/test/metadata_test.dart
@@ -22,15 +22,6 @@ const _emptySourceMetadata =
     '"fileUri":"org-dartlang-app:///web/main.dart","partUris":[]}]}\n'
     '// intentionally empty: package blah has no dart sources';
 
-const _noNullSafetyMetadata =
-    '{"version":"1.0.0","name":"web/main","closureName":"load__web__main",'
-    '"sourceMapUri":"foo/web/main.ddc.js.map",'
-    '"moduleUri":"foo/web/main.ddc.js",'
-    '"libraries":[{"name":"main",'
-    '"importUri":"org-dartlang-app:///web/main.dart",'
-    '"fileUri":"org-dartlang-app:///web/main.dart","partUris":[]}]}\n'
-    '// intentionally empty: package blah has no dart sources';
-
 const _fileUriMetadata =
     '{"version":"1.0.0","name":"web/main","closureName":"load__web__main",'
     '"sourceMapUri":"foo/web/main.ddc.js.map",'
@@ -57,19 +48,6 @@ void main() {
       await provider.libraries,
       contains('org-dartlang-app:///web/main.dart'),
     );
-    expect(await provider.soundNullSafety, isNotNull);
-  });
-
-  test('can parse metadata with no null safety information', () async {
-    final provider = MetadataProvider(
-      'foo.bootstrap.js',
-      FakeAssetReader(metadata: _noNullSafetyMetadata),
-    );
-    expect(
-      await provider.libraries,
-      contains('org-dartlang-app:///web/main.dart'),
-    );
-    expect(await provider.soundNullSafety, true);
   });
 
   test('throws on metadata with absolute import uris', () async {