+++ dwds/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - Replace deprecated JS code `this.__proto__` with `Object.getPrototypeOf(this)`. - [#2500](https://github.com/dart-lang/webdev/pull/2500)
 - Migrate injected client code to `package:web`. - [#2491](https://github.com/dart-lang/webdev/pull/2491)
-- Removed unsound null safety options from classes ModuleMetadata, MetadataProvider & Metadata_test. - [#2427](https://github.com/dart-lang/webdev/issues/2427)
+- Deprecated MetadataProvider's soundNullSafety'. - [#2427](https://github.com/dart-lang/webdev/issues/2427)
 
 ## 24.1.0